nuribom_display_part_test/socket_test/chat/requirements.txt
bidict==0.22.0
call==0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.1
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.0.post0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
pycparser==2.21
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.1
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.12
websocket-client==1.4.2
Werkzeug==2.2.2
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.1
